+++ include/common/ssh.h
@@ -8,10 +8,10 @@
 #include <memory>
 #include <functional>
 
-#include "sftp.h"
 
 struct ssh_session_struct;
 struct ssh_channel_struct;
+struct sftp_session_struct;
 
 namespace util {
     class SSHChannel {
@@ -36,6 +36,39 @@ namespace util {
     private:
         ssh_channel_struct* _channel{};
         int _timeout = 1000;
+    };
+
+    class SFTPSession {
+    public:
+        static std::unique_ptr<SFTPSession> NewSFTPSession(ssh_session_struct* session);
+
+        ~SFTPSession();
+
+        SFTPSession(const SFTPSession&) = delete;
+
+        SFTPSession(SFTPSession&&) = delete;
+
+        void printError() const;
+
+        // remoteFilePath: if you want to store the file in remote working path please set with "./fileName"
+        bool putFile(const std::string& remoteFilePath, bool override, void* data, int size);
+
+        // remoteFilePath: if you want to store the file in remote working path please set with "./fileName"
+        bool putFile(const std::string& remoteFilePath, bool override, const std::string& localFilePath);
+
+        // remoteFilePath: the file name and the relevant path
+        bool getFileToDisk(const std::string& remoteFilePath, const std::string &localFilePath, bool override);
+
+        bool getFileToBuffer(const std::string& remoteFilePath, std::string& buffer);
+
+    protected:
+        SFTPSession() = default;
+
+        bool getFileWithCallback(const std::string& remoteFilePath, const std::function<bool(void* data, int size)>& callback);
+
+    private:
+        sftp_session_struct* _sftp{};
+        ssh_session_struct* _session{};
     };
 
     class SSHSession {
